Add Imu functionality.
The two new commands are:
"i" <addr> where <addr> is the address of the Imu
on the i2c bus.
"j" <addr> <type> where <addr> is as above and <type> is a string
describing the MPU type , e.ig. MPU6050

The Imu objects are kept in an stl map object with the <addr> as the key.